Renamed a few files and added comments
